requirements.txt
aiogram==2.25.1
gspread==6.1.4
google-auth==2.33.0
